separacion de logica
