+++ src/domain/fetchItems.ts
@@ -1,0 +1,5 @@
+import { fetchItemsAPI } from "../infrastructure/itemRepository";
+
+export const fetchItems = async () => {
+  return await fetchItemsAPI();
+};
