+++ src/infrastructure/addItems.ts
@@ -1,0 +1,5 @@
+import { addItemAPI } from "./itemRepository";
+
+export const addItem = async (newItem) => {
+  return await addItemAPI(newItem);
+};
