+++ src/infrastructure/itemRepository.tsx
@@ -1,0 +1,11 @@
+import axios from 'axios';
+
+export const fetchItemsAPI = async () => {
+  const response = await axios.get('https://jsonplaceholder.typicode.com/posts');
+  return response.data;
+};
+
+export const addItemAPI = async (newItem) => {
+  const response = await axios.post('https://jsonplaceholder.typicode.com/posts', newItem);
+  return response.data;
+};
