+++ src/presentation/hooks/useAddItem.ts
@@ -1,0 +1,11 @@
+import { useMutation, useQueryClient } from "react-query";
+import { addItem } from "../../infrastructure/addItems";
+
+export const useAddItem = () => {
+  const queryClient = useQueryClient();
+  return useMutation(addItem, {
+    onSuccess: () => {
+      queryClient.invalidateQueries('items');
+    },
+  });
+};
